LayoutTests/js/dom/modules/missing-exception-check-for-import-expected.txt

Test module import with validateExceptionChecks.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Import should fail because path is not a file.
PASS String(error).substr(0, 11) is "TypeError: "

Importing a module script should fail if the file is not found.
PASS String(error) is "TypeError: Importing a module script failed."

Importing a module script should fail if it does not not a valid JavaScript MIME type.
PASS String(error) is "TypeError: 'text/html' is not a valid JavaScript MIME type."

Importing a simple module script should succeed.
PASS error is null
PASS successfullyParsed is true

TEST COMPLETE
